Adds hooks, Adds filters, Fixes thumb column.
Adds nonce key to main class. Adds nonce methods to main class. Adds assset hooks to header and footer to enqueue scripts and styles inline as efficiently as possible and with the fewest amount of checks. Allows shortcircuit of column content. Adds before and after hooks to all fields according to object type.